Refactort new two places where you can lay your own commands.
